port A1 code to A2
just a sloppy conversion of the A1 code to train for A2. This works
pretty well since both projects are 10-class classification problems, i
basically just swapped out the model and preprocessing using the stuff
from the ipython notebook. This works now for training and can do
minibatches. TODO: validation and test.

Also note that the data is in its own stl10_matlab folder now, not in
the main directory